Create ShowOffersActivity with its RecycleView
I created an activity which takes a RecycleView to show different Offers
to the user. The tests are implemented and should be functional, but i
still have trouble finding how to test the RecycleView with Espresso, so
for now, just the creation of the View (with some dummy offers) is
tested without specifically checking if everything is correct.

I changed some of the names of the Offer. It seems more reasonable to me
to have an author and a title rather than a username and a name, which
are too generic.