Integrate the database for offers
I updated the CreateOfferActivity so that when an offer is created, it send it
to the database. It then show the offer if it worked or else show an error
message.
The ListOfferActivity is also adapted to get data from the database. We still
need to handle error when we can't get data.